# openapi.yaml
openapi: 3.1.0
info:
  title: Biensperience API
  description: |
    # Biensperience Travel Experience Platform API

    Biensperience is a visual travel experience platform that enables users to plan experiences and share their plans with other users who can plan similar experiences. Users can create detailed travel plans with checklists, photos, and hierarchical plan items, then share these experiences to inspire and guide others in planning their own adventures.

    ## Authentication

    The API supports multiple authentication methods:

    ### JWT Authentication
    Most endpoints require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ### OAuth Authentication
    Social login via Facebook, Google, and Twitter is supported through dedicated OAuth endpoints.

    ### API Tokens
    For programmatic access, API tokens can be used instead of JWT tokens.

    ## Rate Limiting

    The API implements comprehensive rate limiting:
    - **Authentication endpoints**: 5 requests per 15 minutes
    - **API endpoints**: 1000 requests per 15 minutes (configurable)
    - **Modification endpoints**: 100 requests per 15 minutes
    - **Collaborator management**: 50 requests per 15 minutes
    - **AI endpoints**: 100 requests per 15 minutes
    - **BienBot endpoints**: 30 requests per 15 minutes (super admins exempt)

    ## Feature Flags

    Certain features require feature flags to be enabled on the user account:

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiTokenAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or OAuth flow
    apiTokenAuth:
      type: apiKey
      in: header
      name: X-API-Token
      description: API token for programmatic access
  schemas:
...
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response data
        message:
          type: string
          description: Success message
      required:
        - success
...
paths:
...
  /users:
    post:
      summary: Create User
      description: Creates a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 3
                inviteCode:
                  type: string
              required:
                - name
                - email
                - password
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Get All Users
      description: Returns a list of all users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalCount:
                        type: integer
                      totalPages:
                        type: integer
                      hasMore:
                        type: boolean
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/search:
    get:
      summary: Search Users
      description: Searches for users by name or email
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
...
  /users/login:
    post:
      summary: User Login
      description: Authenticates a user and returns a JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: JWT token
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /destinations:
    get:
      summary: Get Destinations
      description: Returns a list of destinations
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/searchQuery'
      responses:
        '200':
          description: List of destinations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Destination'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalCount:
                        type: integer
                      totalPages:
                        type: integer
                      hasMore:
                        type: boolean

    post:
      summary: Create Destination
      description: Creates a new destination
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                country:
                  type: string
                  minLength: 1
                description:
                  type: string
              required:
                - name
                - country
      responses:
        '201':
          description: Destination created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /experiences:
    get:
      summary: Get Experiences
      description: Returns a list of experiences
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/searchQuery'
        - $ref: '#/components/parameters/tag'
      responses:
        '200':
          description: List of experiences
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Experience'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalCount:
                        type: integer
                      totalPages:
                        type: integer
                      hasMore:
                        type: boolean

    post:
      summary: Create Experience
      description: Creates a new experience
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                overview:
                  type: string
                destination:
                  type: string
                experience_type:
                  type: array
                  items:
                    type: string
                plan_items:
                  type: array
                  items:
                    $ref: '#/components/schemas/PlanItem'
                visibility:
                  type: string
                  enum: [private, contributors, public]
                  default: public
                difficulty:
                  type: number
                  minimum: 1
                  maximum: 10
                rating:
                  type: number
                  minimum: 0
                  maximum: 5
              required:
                - name
                - destination
      responses:
        '201':
          description: Experience created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Email verification required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /plans:
    get:
      summary: Get User Plans
      description: Returns all plans for the current user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/paginate'
      responses:
        '200':
          description: List of user plans
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Plan'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Plan'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalCount:
                            type: integer
                          totalPages:
                            type: integer
                          hasMore:
                            type: boolean
...
  /photos:
    post:
      summary: Upload Photo
      description: Uploads a single photo file
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, WebP, max 10MB)
              required:
                - image
      responses:
        '201':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Invalid file format or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /search:
    get:
      summary: Global Search
      description: Searches across all collections (destinations, experiences, plans, users)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/searchQuery'
        - $ref: '#/components/parameters/searchTypes'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
  /search/destinations:
    get:
      summary: Search Destinations
      description: Searches destinations by name or country
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/searchQuery'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Destination search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Destination'
  /search/experiences:
    get:
      summary: Search Experiences
      description: Searches experiences by name, overview, or tags
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/searchQuery'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Experience search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Experience'
  /search/plans:
    get:
      summary: Search Plans
      description: Searches plans by experience name or notes
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/searchQuery'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Plan search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan'
  /search/users:
    get:
      summary: Search Users
      description: Searches users by name or email
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/searchQuery'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: User search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /dashboard:
    get:
      summary: Get Dashboard
      description: Returns user's dashboard with stats, recent activity, and upcoming plans
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
...
  /tokens:
    get:
      summary: Get API Tokens
      description: Returns all API tokens for the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of API tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Token'

    post:
      summary: Create API Token
      description: Creates a new API token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                permissions:
                  type: array
                  items:
                    type: string
                    enum: [read, write, admin]
                  default: [read]
                expiresAt:
                  type: string
                  format: date-time
              required:
                - name
      responses:
        '201':
          description: Token created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    description: The actual API token (only shown once)
                  tokenData:
                    $ref: '#/components/schemas/Token'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...